v8: fix mappings synchronization
The prefix list generation was not updated at all when an existing
mapping is seen again. To avoid extra work, create a per-mapping
uint32 which is shared via pointer to the mapping and all prefixes
generated from it. This allows updating the generation for all of
them with one assignment.

Reported-by: Nayef Ghattas <nayef.ghattas@datadoghq.com>