reporter: adapt ReportCountForTrace
In https://github.com/elastic/otel-profiling-agent/pull/90 the signature
for ReportTraceEvent() got changed to allow PID utilization to enrich
traces. This proposal follows this change and also updates
ReportCountForTrace() to use PID as argument.
With this change both API functions are aligned to provide a unified way to
enrich traces with additional metadata.

Signed-off-by: Florian Lehner <florian.lehner@elastic.co>