Add "opaquify" hack
This is necessary in some cases when we need the verifier to _not_ be
able to predict the bounds of a variable, which can cause an explosion
in the amount of state it needs to consider and prevent verification
in under 1M instructions.

See the comment for fuller details.

This also requires setting the Clang target to `bpf` and passing the
arm64/x86 defines manually (rather than relying on `-target` to set
them), because otherwise the inline eBPF assembly doesn't work.