Usermode-side Go custom labels support
If the `CollectCustomLabels` config flag is turned on,
this parses Go binaries to determine their version number, then uses a
hardcoded map of offsets by version to populate the `go_procs` map and
instruct the unwinder to collect custom labels.

In the future, we can automate updating this map, and also fall back
to finding the labels via disassembly when a given version doesn't
have them available. Those features are not yet present in this
commit, however.