reporter: drop project ID
Reporting a project ID originates from a SaaS setup. Currently neither
users of the agent nor users of the reporter package can set the project
ID and so a default value is sent every time.

In the OTel environment some element in OTel ResourceProfile should be
preferred to set such a value.

Signed-off-by: Florian Lehner <florian.lehner@elastic.co>