interpreters, reporter: refactor towards statelessness
Due to legacy reasons, each interpreter kept their own state of
which dynamic metadata should be sent to the reporter. Several
of these caches would never expire, causing caching issues in
the otlp reporter module.

This removes the caching state from all interpreters and pushes
it to the reporter module. A new reporter API call FrameNeeded
is added to query if a specific Frame is in the cache or not.
Not all interpreter modules use the call as all the information
might be available with little overhead. FrameMetadata is also
updated to use the FrameID type for symmetry.

Improved are:
 - reduced memory overhead as per-interpreter caches are removed
 - reporter module can now control which frames need resolving
 - fixes otlp to get the frames re-symbolized if its internal
   lru already forgot about the earlier symbolization information

ref #121