BPF: dotnet: further fixup nibble map reading for verifier
As a follow up to #142, it was discovered that the clang generated
code is still not fully compliant with all supporter kernel verifiers,
and this will be caught in future with PR #205.

Simplify and rearrange the code sequence so that clang chooses
to generate code that is more verifier compliant.