Fix unwinding at syscall on aarch64
Previously, we were using the `return_address` field for two separate
purposes on aarch64:

1. Determining whether we need to subtract 1 from the address during
   later processing, and

2. Determining whether we are in a leaf frame and can therefore trust
   that the current value of x30/LR is indeed the return address.

In most cases, these conditions are the same, but one case they're not
is for syscall return addresses. Linux seems to preserve x30 across
syscalls, so it's still fine to use it to unwind from a frame that's
making a syscall.

Thus we should split this one variable into two, one for each use
case, so we don't artificially fail to unwind in cases where we need
to use LR-based unwinding and are in a syscall frame.