tracehandler: set lifetime for cache
package tracehandler uses a cache to decide wheter frames for traces should be
reported or not. Currently, the lifetime for elements in this cache is
indefinetly, unless the element gets evicted if too many elements are added to
the cache.
Holding back frames for traces can lead to issues with the backend missing
information, if the backend was restarted or there was an issue sending the
frame information for a trace in the first place.

Signed-off-by: Florian Lehner <florian.lehner@elastic.co>