Fix varinfo parsing for luajit local var name lookup
We were decrementing the slot when PC was out of bounds which is wrong
because LuaJIT reuses stack slots for locals in different scopes. This
caused us to hit on local var lookup instead of failing and falling
through to looking at bytecode names.

Add a test case lifted from luajit that shows slots being reused for
different variables and remove old unhelpful test case.