fix: correctly calculate sp for JSEntry Frame layout
Starting v8 8.0.13, the JSEntry frame layout was updated
to accomodate unwinding on Windows ARM64. The changes made
on this PR take into account the stack size for this frame,
which is 160 instead of 16 as multiple callee-saved registers
are pushed onto it before fp and lr.

The fix only works starting v8 9.0.17: the JSEntry frame
layout was updated again to include a valid saved frame pointer.
Prior, the frame pointer (x29) did not point to the previous
{x29, x30} pair on the stack, thus not conforming to the Windows
ARM64 ABI Conventions [1].

[1] https://learn.microsoft.com/en-us/cpp/build/arm64-windows-abi-conventions?view=msvc-160#stack-walking

* interpreter/nodev8/v8.go
* support/ebpf/tracer.ebpf.release.arm64
* support/ebpf/v8_tracer.ebpf.c