Add Process.ExtractAsFile interface
For live processes a direct filename to /proc is returned,
and for coredump cases a temporary file is created with
the data extracted.

This is a last resort method to access the contets and mainly
suitable if the is inspected via 3rd party module/library which
does not support Go ReaderAt interface.