Fix X86 tls handling (for custom labels) (#63)
See the comment for more details -- we were not accounting for the fact that the pointer to the end of the TLS segment might be aligned. Which is strange, because alignment normally affects the beginning of things, not the end of things, but apparently this is how things work.

Also, greatly simplify how we calculate the TLS offset on x86. We don't need to look at the individual sections and deduce how they'll be combined into a segment at runtime; we can just get the information about the segment from the program header directly.