Fix two problems with x86 extraction logic
Problem 1 was the real problem and is was just the compiler using a
register to write a zero instead of an immediate.

Problem 2 was a secondary problem where our simple extraction logic
with symbols didn't match our more complex logic w/o symbols. This
was due to a broken call pattern match utility that didn't properly
detect the AABA call pattern.

Testing done with local shared library. Extraction into proper unit
tests will come in a follow on PR.