pfelf: fix and optimize mmapped section reading
The Section reader elfReader was assigned the SectionReader, not
the backing pfelf.File's elfReader. This caused the dynamic cast
to mmap.ReaderAt always fail in Section.Data().

Fix the assignment to pass the backing elfReader, for the cast to
work correctly. And while at it, remove the SectionReader as it
causes GC pressure and implement the ReadAt() function on the struct
directly. This also fixes the Section to implement io.ReaderAt
itnerface which commit 9788d396 accidentally removed by removing
the embedded io.ReaderAt interface. Add also interace contract
tests in the file.