refactor elfgopclntab to support symbolization (#523)
This refactors elfgopclntab code to be more readable, and implements support for Golang symbolization using mmap backed data.

A building block for #456 to not depend on debug/elf and allow Golang symbolization without excessive memory usage. So the first element in solving the recent memory usage increases and dependency on CGO.