reporter: respect existing elements in dictionary
The helper elements, like stringMap and FuncMap, in setProfile() assumed, that they are the only and always the first that write to protocol global dictionaries. This is not correct and so while the content of these helper elements were appended to the according dictionaries, the references were not correct (as they always started at 0).

Fixes #534

Thanks @fandreuz for reporting this issue.

Signed-off-by: Florian Lehner <florian.lehner@elastic.co>