reporter: extract container.ID from cgroupv2 path
This is a follow up to https://github.com/open-telemetry/opentelemetry-ebpf-profiler/pull/535/.

While the tracer part of the project reports the cgroup v2 path for each sample, the
reporter is expected to report the container ID. The container ID then can be used to
associate the sample to more detailed resource information.

In the context of OTel collector, the container ID then can be used like this:

```
  k8sattributes:
    auth_type: "serviceAccount"
    passthrough: false
    filter:
      node_from_env_var: KUBERNETES_NODE_NAME
    extract:
      metadata:
        - k8s.pod.name
        - k8s.pod.uid
        - k8s.deployment.name
        - k8s.namespace.name
        - k8s.node.name
        - k8s.pod.start_time
        - service.namespace
        - service.name
        - service.version
        - service.instance.id
      labels:
        - tag_name: app.label.component
          key: app.kubernetes.io/component
          from: pod
      otel_annotations: true
    pod_association:
      - sources:
          - from: resource_attribute
            name: container.id
```

As the cgroupv2 path contains further information, that could be beneficial for other
reporters, the extraction of the container ID happens in the reporter.

Signed-off-by: Florian Lehner <florian.lehner@elastic.co>