ebpf: introduce with_debug_output
With this change, if the eBPF profiler is running with `-v` the eBPF debug output will be printed to tracelog:

```
$ sudo ./ebpf-profiler -collection-agent=127.0.0.1:11000 -disable-tls -v
$ sudo bpftool prog tracelog
```

We need to set `-g` to the compile flag for the eBPF objects, as the BTF section is required to store the variable. With this change applied, I was not able to measure a difference in processing the eBPF programs.

If this idea gets accepted, this idea can also be applied to values that are stored in the system_config map. Consequently removing bpf_map_lookup_elem() from this map.

Fixes #468

Signed-off-by: Florian Lehner <florian.lehner@elastic.co>