elfehframe: dynamically detect entry stub (#573)
Earlier we synthesized STOP delta for FDE ending at the entry point. This does not work reliably: 1. in some cases there is no FDE preceding the entry point and 2. often the musl entry code's second portion of code immediately following the CALL which has an FDE that will invalidate the STOP delta.

To better handle the situation this will dynamically match the entry code against known stubs when needed and mark the whole stub with STOP overriding any FDEs.