Remove host.FileID from StackDeltaProvider interface
No one uses this. It was originally envisioned to be used for possibly
implementing alternative provider to fetch the deltas from remove server
using the host.FileID. But this is now superceded by the fact that
pfelf.Reference is passed and the GnuBuildID or other checksums can
be calculated if needed.