pfelf: remove debug/elf helpers
These are legacy cruft and not used in ebpf-profiler. They were
used in the original code. These helpers were also used in the
original symbolizer (which was not published as part of this
tree).

Clean up the code by removing this now unused, and obsolete
helpers. All usage to ELF files should happen via pfelf.File
and not via debug/elf.File.