Allow pushing an "extra" value in Frame in padding
Currently there are 6 bytes of unused padding on the Frame struct.
They are present for alignment purposes, but can be used to ferry
an additional address to userspace if needed.

This enables `_push_with_extra` on the BPF side, as well as the
code to decode this on the userspace side if it is present.